Generalize the Windows-reserved-name handling into shared helpers
The escape/override logic was pasted per consumer: the same Set.kt
code-reading helper copied into three scripts, a one-off regex in the
snapshot test, and mtgish Autogen still lowercasing codes into paths
(its relocate mode would even mkdirs definitions/con/ again).

- scripts/set_dirs.py: shared dir<->code mapping module; the three
  reprint/printing scripts import it instead of carrying copies.
- GoldenSnapshot.fileSafe: the reserved-name escape lives with the
  golden-file utility, not the one test that happened to need it.
- mtgish Support.setDirSegment: resolves a set code to its actual
  definitions/ directory (Set.kt is authoritative), underscore-escape
  fallback for unscaffolded reserved codes. All four Autogen path
  sites use it, closing the coverage-generate --set CON gap.
- windowsSafeFileName/fromWindowsSafeFileName round-trip pair applied
  to fixture and Scryfall-cache filenames minted from set codes.
