+++ scripts/generate-reprints.py
@@ -34,6 +34,8 @@ import time
 from collections import defaultdict
 from pathlib import Path
 
+from set_dirs import dir_for_codes, scaffolded_set_codes, set_dir_codes
+
 REPO_ROOT = Path(__file__).resolve().parent.parent
 DEFINITIONS_ROOT = REPO_ROOT / "mtg-sets/src/main/kotlin/com/wingedsheep/mtg/sets/definitions"
 PRINTINGS_CACHE = Path.home() / ".cache" / "scryfall" / "printings"
@@ -67,35 +69,6 @@ def pascal(name: str) -> str:
     if pas and pas[0].isdigit():
         pas = "N" + pas
     return pas
-
-
-SET_CODE_RE = re.compile(r'override\s+val\s+code\s*=\s*"([^"]+)"')
-
-
-def set_dir_codes() -> dict[str, str]:
-    """Map each definitions/<dir> to its lowercase set code, read from the dir's *Set.kt.
-
-    The directory name usually equals the code, but can't always: `con` is a reserved
-    filename on Windows, so Conflux lives in `definitions/conflux/`.
-    """
-    codes: dict[str, str] = {}
-    if not DEFINITIONS_ROOT.is_dir():
-        return codes
-    for d in sorted(DEFINITIONS_ROOT.iterdir()):
-        if not d.is_dir():
-            continue
-        code = None
-        for set_kt in sorted(d.glob("*Set.kt")):
-            m = SET_CODE_RE.search(set_kt.read_text(encoding="utf-8"))
-            if m:
-                code = m.group(1).lower()
-                break
-        codes[d.name] = code or d.name
-    return codes
-
-
-def scaffolded_sets() -> set[str]:
-    return set(set_dir_codes().values())
 
 
 def scan_definitions() -> tuple[dict[str, str], dict[str, set[str]]]:
@@ -219,8 +192,8 @@ def main() -> int:
     args = ap.parse_args()
     only = args.set.lower() if args.set else None
 
-    scaffolded = scaffolded_sets()
-    dir_for = {code: d for d, code in set_dir_codes().items()}
+    scaffolded = scaffolded_set_codes()
+    dir_for = dir_for_codes()
     canonical, reprints = scan_definitions()
 
     written = 0
