+++ scripts/missing-reprints.py
@@ -38,6 +38,8 @@ from collections import defaultdict
 from dataclasses import dataclass
 from pathlib import Path
 
+from set_dirs import scaffolded_set_codes, set_dir_codes
+
 REPO_ROOT = Path(__file__).resolve().parent.parent
 DEFINITIONS_ROOT = REPO_ROOT / "mtg-sets/src/main/kotlin/com/wingedsheep/mtg/sets/definitions"
 CACHE_ROOT = Path.home() / ".cache" / "scryfall" / "printings"
@@ -139,35 +141,6 @@ def fetch_printings(card_name: str) -> list[Printing]:
     return printings
 
 
-SET_CODE_RE = re.compile(r'override\s+val\s+code\s*=\s*"([^"]+)"')
-
-
-def set_dir_codes() -> dict[str, str]:
-    """Map each definitions/<dir> to its lowercase set code, read from the dir's *Set.kt.
-
-    The directory name usually equals the code, but can't always: `con` is a reserved
-    filename on Windows, so Conflux lives in `definitions/conflux/`.
-    """
-    codes: dict[str, str] = {}
-    if not DEFINITIONS_ROOT.is_dir():
-        return codes
-    for d in sorted(DEFINITIONS_ROOT.iterdir()):
-        if not d.is_dir():
-            continue
-        code = None
-        for set_kt in sorted(d.glob("*Set.kt")):
-            m = SET_CODE_RE.search(set_kt.read_text(encoding="utf-8"))
-            if m:
-                code = m.group(1).lower()
-                break
-        codes[d.name] = code or d.name
-    return codes
-
-
-def scaffolded_sets() -> set[str]:
-    return set(set_dir_codes().values())
-
-
 def scan_definitions() -> tuple[dict[str, str], dict[str, set[str]]]:
     """Return (canonical[name]->set_code, reprints[name]->{set_codes})."""
     canonical: dict[str, str] = {}
@@ -204,7 +177,7 @@ def main() -> int:
     parser.add_argument("--limit", type=int, default=0, help="cap number of Scryfall fetches")
     args = parser.parse_args()
 
-    scaffolded = scaffolded_sets()
+    scaffolded = scaffolded_set_codes()
     canonical, reprints = scan_definitions()
 
     only_set = args.set.lower() if args.set else None
